Шаблонизация проекта: вариант второй
